Eric/pla 845 dirty check parquet column before write (#6)
* Add dirty checking to columns

* rm test file


Former-commit-id: ea61f9e8f6e901cdc78b375add3db5190185e975

+++ src/main.rs
@@ -29,21 +29,23 @@ fn main() {
                 column.remove_value(email_to_remove.clone(), email_to_replace.clone());
             }
         }
-        // Make a backup before writing.
-        let mut backup_path = file_path.path();
-        backup_path.set_extension("parquet.backup");
-        copy(file_path.path(), backup_path.clone()).unwrap();
 
-        // Write the final file, if we fail then restore the backup.
-        match file.write_to(file_path.path().as_ref()) {
-            Ok(_) => {
-                remove_file(backup_path).unwrap();
-            }
-            Err(err) => {
-                copy(&backup_path, file_path.path()).unwrap();
-                remove_file(&backup_path).unwrap();
-                panic!("Error while writing file: {:?}", err)
-            }
-        };
+        if file.is_dirty() {
+            // Make a backup before writing.
+            let mut backup_path = file_path.path();
+            backup_path.set_extension("parquet.backup");
+            copy(file_path.path(), backup_path.clone()).unwrap();
+            // Write the final file, if we fail then restore the backup.
+            match file.write_to(file_path.path().as_ref()) {
+                Ok(_) => {
+                    remove_file(backup_path).unwrap();
+                }
+                Err(err) => {
+                    copy(&backup_path, file_path.path()).unwrap();
+                    remove_file(&backup_path).unwrap();
+                    panic!("Error while writing file: {:?}", err)
+                }
+            };
+        }
     }
 }
